Refactored type helpers into multiple files
