+++ packages/realm/src/type-helpers/LinkingObjects.ts
@@ -1,0 +1,33 @@
+////////////////////////////////////////////////////////////////////////////
+//
+// Copyright 2024 Realm Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////
+
+import { assert, binding } from "../internal";
+import { defaultToBinding } from "./default";
+import { TypeHelpers, TypeOptions } from "./types";
+
+export function createLinkingObjectsTypeHelpers({ objectType, getClassHelpers }: TypeOptions): TypeHelpers {
+  assert(objectType);
+  const { wrapObject } = getClassHelpers(objectType);
+  return {
+    toBinding: defaultToBinding,
+    fromBinding(value) {
+      assert.instanceOf(value, binding.Obj);
+      return wrapObject(value);
+    },
+  };
+}
